+++ src/loaders/types.js
@@ -7,8 +7,11 @@ import ClustersJsonAsMatrixZarrLoader from './ClustersJsonAsMatrixZarrLoader';
 import RasterJsonLoader from './RasterJsonLoader';
 import OmeZarrLoader from './OmeZarrLoader';
 import CellSetsJsonLoader from './CellSetsJsonLoader';
+import MoleculesJsonLoader from './MoleculesJsonLoader';
 import AnnDataLoaders from './anndata-loaders';
 import GenomicProfilesZarrLoader from './GenomicProfilesZarrLoader';
+import CellsGeoJsonLoader from './CellsGeoJsonLoader';
+
 import { AnnDataSource, ZarrDataSource, JsonSource } from './data-sources';
 
 export const fileTypeToLoaderAndSource = {
@@ -16,7 +19,7 @@ export const fileTypeToLoaderAndSource = {
   [FileType.CLUSTERS_JSON]: [JsonSource, ClustersJsonAsMatrixZarrLoader],
   [FileType.GENES_JSON]: [JsonSource, GenesJsonAsMatrixZarrLoader],
   [FileType.CELLS_JSON]: [JsonSource, JsonLoader],
-  [FileType.MOLECULES_JSON]: [JsonSource, JsonLoader],
+  [FileType.MOLECULES_JSON]: [JsonSource, MoleculesJsonLoader],
   [FileType.NEIGHBORHOODS_JSON]: [JsonSource, JsonLoader],
   [FileType.RASTER_JSON]: [JsonSource, RasterJsonLoader],
   [FileType.RASTER_OME_ZARR]: [ZarrDataSource, OmeZarrLoader],
@@ -24,7 +27,9 @@ export const fileTypeToLoaderAndSource = {
   [FileType.ANNDATA_CELL_SETS_ZARR]: [AnnDataSource, AnnDataLoaders.CellSetsZarrLoader],
   [FileType.ANNDATA_CELLS_ZARR]: [AnnDataSource, AnnDataLoaders.CellsZarrLoader],
   [FileType.ANNDATA_EXPRESSION_MATRIX_ZARR]: [AnnDataSource, AnnDataLoaders.MatrixZarrLoader],
+  [FileType.ANNDATA_MOLECULES_ZARR]: [AnnDataSource, AnnDataLoaders.MoleculesZarrLoader],
   [FileType.GENOMIC_PROFILES_ZARR]: [ZarrDataSource, GenomicProfilesZarrLoader],
+  [FileType.CELLS_GEOJSON]: [JsonSource, CellsGeoJsonLoader],
 };
 
 export function getSourceAndLoaderFromFileType(type) {