+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 - Added the `scoreName` property to the view config schema for the `anndata-cell-sets.zarr` file type.
 - Added a new documentation site.
 - Added the `./create-release.sh` Bash script to automate some steps of the release process.
+- Added coordination types for selection of molecules.
+- Added `cells.geojson` loader to support the Baysor cell segmentation polygon format.
+- Added `anndata-molecules.zarr` loader to support the Baysor molecules format.
+- Added support for multiple gene selection via `shift+click`.
+- Added a `use3d` option for `spatialMoleculesLayer` to enable a point-cloud rendering of 3D molecules data.
 
 ### Changed
 - Fix selection issue for bitmasks in external applications.
@@ -14,6 +19,7 @@
 - Bump dependency versions based on `npm audit` and dependabot pull requests.
 - Updated Heatmap `layerFilter` function to reflect [changes between deck.gl 8.5 and 8.6 ](https://deck.gl/docs/upgrade-guide#layer-filtering)
 - Updated R package URLs (to reflect repo name change from `vitessce-r` to `vitessceR`).
+- Changed the `molecules.json` loader class to reflect the new internal data structure for molecules.
 
 
 ## [1.1.17](https://www.npmjs.com/package/vitessce/v/1.1.17) - 2021-11-04